+++ docs/website/esbuild.config.mjs
@@ -7,8 +7,10 @@ const DEV = process.env.NODE_ENV === 'DEV';
 const jsFolder = DEV ? 'lib/site' : 'build/site';
 
 const tsEntrypoints = [
+  './src/site/ts/utils/active-element.ts',
   './src/site/ts/components/my-counter.ts',
   './src/site/ts/components/test-component.ts',
+  './src/site/ts/scripts/*.ts',
   './src/site/ts/ssr-utils/lit-hydrate-support.ts',
   './src/site/ts/ssr-utils/is-land.ts'
 ];