+++ docs/website/site.eleventy.cjs
@@ -2,6 +2,7 @@ const litPlugin = require('@lit-labs/eleventy-plugin-lit');
 const eleventyNavigationPlugin = require('@11ty/eleventy-navigation');
 const markdownItAnchor = require('markdown-it-anchor');
 const markdownIt = require('markdown-it');
+const markdownItAttrs = require('markdown-it-attrs')
 const syntaxHighlight = require('@11ty/eleventy-plugin-syntaxhighlight');
 const slugify = require('slugify');
 const htmlMinifier = require('html-minifier');
@@ -60,7 +61,8 @@ module.exports = function(eleventyConfig) {
       permalinkClass: 'direct-link',
       permalinkSymbol: '#',
       level: [1, 2, 3, 4],
-    });
+    })
+    .use(markdownItAttrs);
 
   eleventyConfig.setBrowserSyncConfig({
     notify: true,
@@ -71,7 +73,8 @@ module.exports = function(eleventyConfig) {
 
   eleventyConfig
     .addPassthroughCopy({ './src/shared/assets': 'assets' })
-    .addPassthroughCopy('./src/site/assets');
+    .addPassthroughCopy('./src/site/assets')
+    .addPassthroughCopy({ './../../packages/tokens/src/themes/sanoma-learning': `styles/slds-sanoma-learning` });
 
   const NOT_FOUND_PATH = `${outputFolder}/site/404.html`;