+++ docs/website/src/site/ts/utils/active-element.ts
@@ -1,0 +1,3 @@
+export const getActiveElement = (el: Node): Element | null => {
+  return (el.getRootNode() as Document).activeElement;
+};