+++ packages/core/src/radio-group/radio-group.ts
@@ -2,7 +2,8 @@ import type { CSSResultGroup, PropertyValues, TemplateResult } from 'lit';
 import { FormControlMixin } from '@open-wc/form-control';
 import { LitElement, html } from 'lit';
 import { property, queryAssignedNodes } from 'lit/decorators.js';
-import { RovingTabindexController } from '../utils/controllers/roving-tabindex.js';
+import { EventsController, RovingTabindexController } from '../utils/controllers/index.js';
+import { HintMixin } from '../utils/form-control/index.js';
 import { Radio } from './radio.js';
 import styles from './radio-group.scss.js';
 
@@ -12,53 +13,53 @@ const OBSERVER_OPTIONS: MutationObserverInit = {
   subtree: true
 };
 
-export class RadioGroup extends FormControlMixin(LitElement) {
+export class RadioGroup extends FormControlMixin(HintMixin(LitElement)) {
   /** @private */
   static override styles: CSSResultGroup = styles;
 
-  /** Observe the state of the radios. */
-  #observer?: MutationObserver;
+  /** Events controller. */
+  #events = new EventsController(this);
 
   #rovingTabindexController = new RovingTabindexController<Radio>(this, {
     focusInIndex: (elements: Radio[]) => {
       return elements.findIndex(el => {
-        return this.selected ? !el.disabled && el.value === this.selected : !el.disabled;
+        return this.value ? !el.disabled && el.value === this.value : !el.disabled;
       });
     },
     elementEnterAction: (el: Radio) => {
-      this.selected = el.value;
+      this.value = el.value;
     },
     elements: () => this.buttons,
     isFocusableElement: (el: Radio) => !el.disabled
   });
 
-  get buttons(): Radio[] {
-    return this.defaultNodes?.filter((node): node is Radio => node instanceof Radio) || [];
-  }
+  /** Observe the state of the radios. */
+  #observer?: MutationObserver;
 
   /** The assigned nodes. */
   @queryAssignedNodes() defaultNodes?: Node[];
-
-  /** Whether all the radio's in the group are disabled. */
-  @property({ type: Boolean, reflect: true }) disabled = false;
 
   /** The orientation of the radio's in the group. */
   @property({ type: String, reflect: true }) orientation: 'horizontal' | 'vertical' = 'vertical';
 
   /** The value of the selected radio. */
-  @property() selected = '';
+  @property() value?: string;
+
+  get buttons(): Radio[] {
+    return this.defaultNodes?.filter((node): node is Radio => node instanceof Radio) || [];
+  }
 
   override connectedCallback(): void {
     super.connectedCallback();
 
     this.internals.role = 'radiogroup';
 
+    this.#events.listen(this, 'click', this.#onClick);
+
     this.#observer = new MutationObserver(mutationList => {
       mutationList.forEach(mutation => {
         if (mutation.attributeName === 'checked' && mutation.oldValue === null) {
-          this.selected = (mutation.target as Radio).value;
-
-          this.#updateSelected(this.selected);
+          this.#updateSelected((mutation.target as Radio).value);
         }
       });
     });
@@ -75,20 +76,31 @@ export class RadioGroup extends FormControlMixin(LitElement) {
   override updated(changes: PropertyValues<this>): void {
     super.updated(changes);
 
-    if (changes.has('selected')) {
-      this.#updateSelected(this.selected);
+    if (changes.has('value')) {
+      this.#updateSelected(this.value);
     }
   }
 
   override render(): TemplateResult {
-    return html`<slot @slotchange=${() => this.#rovingTabindexController.clearElementCache()}></slot>`;
+    return html`
+      <div class="wrapper">
+        <slot @slotchange=${() => this.#rovingTabindexController.clearElementCache()}></slot>
+      </div>
+      ${this.renderHint()}
+    `;
   }
 
   override focus(): void {
     this.#rovingTabindexController.focus();
   }
 
-  #updateSelected(value: string): void {
+  #onClick(event: Event): void {
+    event.preventDefault();
+
+    this.focus();
+  }
+
+  #updateSelected(value?: string): void {
     this.#observer?.disconnect();
 
     this.buttons?.forEach(button => {