+++ packages/core/src/textarea/register.ts
@@ -1,0 +1,9 @@
+import { Textarea } from './textarea.js';
+
+customElements.define('sl-textarea', Textarea);
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'sl-textarea': Textarea;
+  }
+}