+++ packages/core/src/tooltip/tooltip-directive.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-non-null-assertion */
 import type { DirectiveParameters, ElementPart } from 'lit/directive.js';
 import { Directive, directive } from 'lit/directive.js';
 import { render } from 'lit';