+++ packages/grid/src/column.ts
@@ -1,12 +1,22 @@
 import type { TemplateResult } from 'lit';
 import type { Grid } from './grid.js';
-import { getNameByPath, getValueByPath } from '@sanomalearning/slds-core/utils';
+import { dasherize, getNameByPath, getValueByPath } from '@sanomalearning/slds-core/utils';
 import { LitElement, html } from 'lit';
 import { property } from 'lit/decorators.js';
+import { ifDefined } from 'lit/directives/if-defined.js';
 
-export type GridColumnRenderer<T> = (model: T) => TemplateResult;
+export type GridColumnHeaderRenderer = () => TemplateResult;
 
-export class GridColumn<T extends { [x: string]: unknown } = Record<string, unknown>> extends LitElement {
+export type GridColumnDataRenderer<T> = (model: T) => TemplateResult;
+
+export type GridColumnParts<T> = (model: T) => string | undefined;
+
+export class GridColumn<T extends Record<string, unknown> = Record<string, unknown>> extends LitElement {
+  #width?: number;
+
+  /** The alignment of the content within the column. */
+  @property() align: 'start' | 'center' | 'end' = 'start';
+
   /**
    * Automatically sets the width of the column based on the column contents when this is set to `true`.
    *
@@ -36,19 +46,29 @@ export class GridColumn<T extends { [x: string]: unknown } = Record<string, unkn
   @property({ type: Number }) grow = 1;
 
   /** The label for the column header. */
-  @property() header?: string;
+  @property() header?: string | GridColumnHeaderRenderer;
 
   /** The path to the value for this column. */
   @property() path?: string;
 
+  /** Custom parts to be set on the `<td>` so it can be styled externally. */
+  @property() parts?: string | GridColumnParts<T>;
+
   /** Renderer function for the column value of each cell. */
-  @property({ attribute: false }) renderer?: GridColumnRenderer<T>;
+  @property({ attribute: false }) renderer?: GridColumnDataRenderer<T>;
 
   /** Whether this column is sticky when the user scrolls horizontally. */
   @property({ type: Boolean, reflect: true }) sticky?: boolean;
 
+  set width(value: number | undefined) {
+    this.#width = value;
+  }
+
   /** Width of the cells for this column in pixels. */
-  @property() width?: number;
+  @property()
+  get width(): number | undefined {
+    return this.#width;
+  }
 
   override connectedCallback(): void {
     super.connectedCallback();
@@ -64,10 +84,22 @@ export class GridColumn<T extends { [x: string]: unknown } = Record<string, unkn
   }
 
   renderData(item: T): TemplateResult {
-    if (this.renderer) {
-      return html`<td>${this.renderer(item)}</td>`;
-    } else {
-      return html`<td>${this.path ? getValueByPath(item, this.path) : 'No path set'}</td>`;
+    let parts;
+
+    if (typeof this.parts === 'string') {
+      parts = this.parts;
+    } else if (typeof this.parts === 'function') {
+      parts = this.parts(item);
     }
+
+    if (this.path) {
+      parts = `${dasherize(this.path.replaceAll('.', '-'))} ${parts || ''}`.trim();
+    }
+
+    return html`
+      <td part=${ifDefined(parts)}>
+        ${this.renderer ? this.renderer(item) : this.path ? getValueByPath(item, this.path) : 'No path set'}
+      </td>
+    `;
   }
 }