+++ packages/grid/src/grid.ts
@@ -1,18 +1,37 @@
 import type { CSSResultGroup, PropertyValues, TemplateResult } from 'lit';
 import type { GridSorter, GridSorterChange } from './sorter.js';
-import type { DataSource, DataSourceSortDirection } from '@sanomalearning/slds-core/utils/data-source';
-import { ArrayDataSource } from '@sanomalearning/slds-core/utils/data-source';
+import type { GridFilter, GridFilterChange } from './filter.js';
+import type { EventEmitter } from '@sanomalearning/slds-core/utils/decorators';
+import type {
+  DataSource,
+  DataSourceFilterValue,
+  DataSourceSortDirection
+} from '@sanomalearning/slds-core/utils/data-source';
 import { virtualize } from '@lit-labs/virtualizer/virtualize.js';
 import { SelectionController } from '@sanomalearning/slds-core/utils/controllers';
+import { ArrayDataSource } from '@sanomalearning/slds-core/utils/data-source';
+import { event } from '@sanomalearning/slds-core/utils/decorators';
 import { LitElement, html } from 'lit';
 import { property, state } from 'lit/decorators.js';
 import { classMap } from 'lit/directives/class-map.js';
 import styles from './grid.scss.js';
 import { GridColumn } from './column.js';
+import { GridColumnGroup } from './column-group.js';
 
-export class Grid<T extends { [x: string]: unknown } = Record<string, unknown>> extends LitElement {
+export class GridActiveItemChangeEvent<T> extends Event {
+  constructor(public readonly item: T, public readonly relatedEvent: Event | null) {
+    super('sl-active-item-change', { bubbles: true, composed: true });
+  }
+}
+
+export type GridItemParts<T> = (model: T) => string | undefined;
+
+export class Grid<T extends Record<string, unknown> = Record<string, unknown>> extends LitElement {
   /** @private */
   static override styles: CSSResultGroup = styles;
+
+  /** The filters for this grid. */
+  #filters: GridFilter[] = [];
 
   /** Flag for calculating the column widths only once. */
   #initialColumnWidthsCalculated = false;
@@ -23,6 +42,12 @@ export class Grid<T extends { [x: string]: unknown } = Record<string, unknown>> 
   /** Selection manager. */
   readonly selection = new SelectionController<T>(this);
 
+  /** The active item in the grid. */
+  @state() activeItem?: T;
+
+  /** Emits when the active item changes */
+  @event() activeItemChange!: EventEmitter<GridActiveItemChangeEvent<T>>;
+
   /** The columns in the grid. */
   @state() columns: Array<GridColumn<T>> = [];
 
@@ -30,13 +55,19 @@ export class Grid<T extends { [x: string]: unknown } = Record<string, unknown>> 
   @property({ attribute: false }) dataSource?: DataSource<T>;
 
   /** An array of items to be displayed in the grid. */
-  @property() items?: T[];
+  @property({ type: Array }) items?: T[];
+
+  /** Custom parts to be set on the `<tr>` so it can be styled externally. */
+  @property({ attribute: false }) itemParts?: GridItemParts<T>;
 
   /** Hide the border around the grid when true. */
   @property({ type: Boolean, reflect: true, attribute: 'no-border' }) noBorder?: boolean;
 
   /** Hides the border between rows when true. */
   @property({ type: Boolean, reflect: true, attribute: 'no-row-border' }) noRowBorder?: boolean;
+
+  /** Uses alternating background colors for the rows when set. */
+  @property({ type: Boolean, reflect: true }) striped?: boolean;
 
   override willUpdate(changes: PropertyValues<this>): void {
     if (changes.has('items')) {
@@ -54,7 +85,7 @@ export class Grid<T extends { [x: string]: unknown } = Record<string, unknown>> 
     super.updated(changes);
 
     if (changes.has('dataSource') && this.dataSource) {
-      this.dataSource?.addEventListener('update', () => (this.selection.size = this.dataSource?.size ?? 0));
+      this.dataSource?.addEventListener('sl-update', () => (this.selection.size = this.dataSource?.size ?? 0));
     }
   }
 
@@ -62,33 +93,21 @@ export class Grid<T extends { [x: string]: unknown } = Record<string, unknown>> 
     return html`
       <slot @slotchange=${this.#onSlotchange} style="display:none"></slot>
       <style>
-        ${this.columns.map((col, index) => {
-          return `
-            :where(td, th):nth-child(${index + 1}) {
-              flex-grow: ${col.grow};
-              width: ${col.width || '100'}px;
-              ${
-                col.sticky
-                  ? `
-                    left: ${this.#getStickyColumnOffset(index)}px;
-                    position: sticky;
-                  `
-                  : ''
-              }
-            }
-          `;
-        })}
+        ${this.renderStyles()}
       </style>
       <table>
-        <thead @directionChange=${this.#onDirectionChange} @sorterChange=${this.#onSorterChange}>
-          <tr>
-            ${this.columns.map(col => col.renderHeader())}
-          </tr>
+        <thead
+          @sl-direction-change=${this.#onDirectionChange}
+          @sl-filter-change=${this.#onFilterChange}
+          @sl-filter-value-change=${this.#onFilterValueChange}
+          @sl-sorter-change=${this.#onSorterChange}
+        >
+          ${this.renderHeader()}
         </thead>
         <tbody @visibilityChanged=${this.#onVisibilityChanged}>
           ${virtualize({
             items: this.dataSource?.items,
-            renderItem: item => this.renderItem(item)
+            renderItem: (item, index) => this.renderItem(item, index)
           })}
         </tbody>
         <tfoot></tfoot>
@@ -96,12 +115,73 @@ export class Grid<T extends { [x: string]: unknown } = Record<string, unknown>> 
     `;
   }
 
-  renderItem(item: T): TemplateResult {
-    const selected = this.selection.isSelected(item);
+  renderStyles(): TemplateResult {
+    const rows = this.#getHeaderRows(this.columns);
 
     return html`
-      <tr class=${classMap({ selected })} part="row ${selected ? 'selected' : ''}">
-        ${this.columns.map(col => col.renderData(item))}
+      ${rows.slice(0, -1).map((row, rowIndex) => {
+        return row.map((col, colIndex) => {
+          return `
+            thead tr:nth-child(${rowIndex + 1}) th:nth-child(${colIndex + 1}) {
+              flex-grow: ${(col as GridColumnGroup<T>).columns.length};
+              justify-content: ${col.align};
+              width: ${col.width || '100'}px;
+            }
+          `;
+        });
+      })}
+      ${rows[rows.length - 1].map((col, index) => {
+        return `
+          :where(td, thead tr:last-of-type th):nth-child(${index + 1}) {
+            flex-grow: ${col.grow};
+            justify-content: ${col.align};
+            width: ${col.width || '100'}px;
+            ${
+              col.sticky
+                ? `
+                  left: ${this.#getStickyColumnOffset(index)}px;
+                  position: sticky;
+                `
+                : ''
+            }
+          }
+        `;
+      })}
+    `;
+  }
+
+  renderHeader(): TemplateResult {
+    const rows = this.#getHeaderRows(this.columns);
+
+    return html`
+      ${rows.map(
+        row =>
+          html`
+            <tr>
+              ${row.map(col => col.renderHeader())}
+            </tr>
+          `
+      )}
+    `;
+  }
+
+  renderItem(item: T, index: number): TemplateResult {
+    const rows = this.#getHeaderRows(this.columns),
+      selected = this.selection.isSelected(item),
+      parts = [
+        'row',
+        index % 2 === 0 ? 'odd' : 'even',
+        ...(selected ? ['selected'] : []),
+        ...(this.itemParts?.(item)?.split(' ') || [])
+      ];
+
+    return html`
+      <tr
+        @click=${(event: Event) => this.#onClickRow(event, item)}
+        class=${classMap({ selected })}
+        part=${parts.join(' ')}
+      >
+        ${rows[rows.length - 1].map(col => col.renderData(item))}
       </tr>
     `;
   }
@@ -111,7 +191,9 @@ export class Grid<T extends { [x: string]: unknown } = Record<string, unknown>> 
     // Do not remove, this is needed; not sure why
     await this.updateComplete;
 
-    this.columns
+    const rows = this.#getHeaderRows(this.columns);
+
+    rows[rows.length - 1]
       .filter(col => !col.hidden && col.autoWidth)
       .forEach(col => {
         const index = this.columns.indexOf(col),
@@ -132,10 +214,27 @@ export class Grid<T extends { [x: string]: unknown } = Record<string, unknown>> 
     this.requestUpdate('columns');
   }
 
+  #onClickRow(event: Event, item: T): void {
+    this.activeItem = item;
+    this.activeItemChange.emit(new GridActiveItemChangeEvent(this.activeItem, event));
+  }
+
   #onDirectionChange({ target }: CustomEvent<DataSourceSortDirection | undefined> & { target: GridSorter }): void {
     this.#sorters.filter(sorter => sorter !== target).forEach(sorter => sorter.reset());
 
     this.#applySorters();
+  }
+
+  #onFilterChange({ detail, target }: CustomEvent<GridFilterChange> & { target: GridFilter }): void {
+    if (detail === 'added') {
+      this.#filters = [...this.#filters, target];
+    } else {
+      this.#filters = this.#filters.filter(filter => filter !== target);
+    }
+  }
+
+  #onFilterValueChange(): void {
+    this.#applyFilters();
   }
 
   #onSlotchange(event: Event & { target: HTMLSlotElement }): void {
@@ -163,21 +262,45 @@ export class Grid<T extends { [x: string]: unknown } = Record<string, unknown>> 
     }
   }
 
+  #applyFilters(): void {
+    if (!this.dataSource) {
+      return;
+    }
+
+    const filterValues: DataSourceFilterValue[] = this.#filters
+      .filter(filter => !!filter.value)
+      .map(filter => ({ path: filter.column.path || '', value: filter.value }));
+
+    this.dataSource.filterValues = filterValues;
+    this.dataSource.update();
+    this.requestUpdate();
+  }
+
   #applySorters(): void {
     if (!this.dataSource) {
       return;
     }
 
-    const { direction, path } = this.#sorters.find(sorter => !!sorter.direction) || {};
+    const { column, direction } = this.#sorters.find(sorter => !!sorter.direction) || {};
 
-    if (direction && path) {
-      this.dataSource.sortValue = { path, direction };
+    if (column?.path && direction) {
+      this.dataSource.sortValue = { path: column.path, direction };
     } else {
       this.dataSource.sortValue = undefined;
     }
 
     this.dataSource.update();
     this.requestUpdate();
+  }
+
+  #getHeaderRows(columns: Array<GridColumn<T>>): Array<Array<GridColumn<T>>> {
+    const children = columns
+      .filter((col): col is GridColumnGroup<T> => col instanceof GridColumnGroup)
+      .reduce((acc: Array<Array<GridColumn<T>>>, cur) => {
+        return [...acc, ...this.#getHeaderRows(cur.columns)];
+      }, []);
+
+    return children.length ? [[...columns], children.flat(2)] : [[...columns]];
   }
 
   #getStickyColumnOffset(index: number): number {