+++ tools/storybook-addon-themes/preset.js
@@ -9,6 +9,7 @@ function previewHead(head) {
       body { 
         background: var(--sl-body-background, #fff);
         color: var(--sl-body-foreground, #333);
+        font: var(--sl-text-body-md-normal);
       }
     </style>
   `;