+++ docs/website/src/site/categories/whats-new.md
@@ -3,5 +3,5 @@ title: What's new
 eleventyNavigation:
   parent: Pages
   key: What's new
-  order: 1
+  order: 2
 ---