+++ packages/grid/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @sanomalearning/slds-grid
 
+## 0.0.2
+
+### Patch Changes
+
+- [#126](https://github.com/sanomalearning/design-system/pull/126) [`1a60a85`](https://github.com/sanomalearning/design-system/commit/1a60a85439787511d3b80beffd99c5a3c6505510) - Add filter column
+
+- [#120](https://github.com/sanomalearning/design-system/pull/120) [`b98a1dc`](https://github.com/sanomalearning/design-system/commit/b98a1dcc01daca4139084f014d4e7cf1b7b9360f) - Add support for striped rows
+
+- [#121](https://github.com/sanomalearning/design-system/pull/121) [`e5370aa`](https://github.com/sanomalearning/design-system/commit/e5370aad42911e8e669c1f53b8f059e09c92acfb) - Add column groups
+
+- [#109](https://github.com/sanomalearning/design-system/pull/109) [`7adf12a`](https://github.com/sanomalearning/design-system/commit/7adf12ad5e0eebe0e7d3b8bc3dabf15ed992b759) - Add sortable columns
+
+- [#127](https://github.com/sanomalearning/design-system/pull/127) [`c816c0a`](https://github.com/sanomalearning/design-system/commit/c816c0ad12707c8361d82d24fba8640eb4608671) - Upgrade to latest `@lit-labs/virtualizer`
+
+- [#120](https://github.com/sanomalearning/design-system/pull/120) [`b98a1dc`](https://github.com/sanomalearning/design-system/commit/b98a1dcc01daca4139084f014d4e7cf1b7b9360f) - Add ability to customize the CSS Shadow Parts of the columns & rows
+
+- [#122](https://github.com/sanomalearning/design-system/pull/122) [`b0b316b`](https://github.com/sanomalearning/design-system/commit/b0b316b9bac800908e9cd88876b850cec39233d1) - Fix autoSelect not working
+
+- [#105](https://github.com/sanomalearning/design-system/pull/105) [`8eca8b5`](https://github.com/sanomalearning/design-system/commit/8eca8b58b9e64ec222299f0a748e4f651775c305) - Add selection column to grid
+
+- [#125](https://github.com/sanomalearning/design-system/pull/125) [`00262de`](https://github.com/sanomalearning/design-system/commit/00262de29be9afa6b035f37bc5535767267b0a6d) - Simplify styling by using the path property for the part attribute
+
+- [#120](https://github.com/sanomalearning/design-system/pull/120) [`b98a1dc`](https://github.com/sanomalearning/design-system/commit/b98a1dcc01daca4139084f014d4e7cf1b7b9360f) - Add support for column content alignment
+
+- Updated dependencies [[`b0b316b`](https://github.com/sanomalearning/design-system/commit/b0b316b9bac800908e9cd88876b850cec39233d1), [`e5b9482`](https://github.com/sanomalearning/design-system/commit/e5b9482896369b2c7ddecd2320f359231b53fb36), [`b0b316b`](https://github.com/sanomalearning/design-system/commit/b0b316b9bac800908e9cd88876b850cec39233d1), [`ac2ba48`](https://github.com/sanomalearning/design-system/commit/ac2ba48b404a9d1e2576b9a97d323feec3176738), [`cd361f9`](https://github.com/sanomalearning/design-system/commit/cd361f9046127d0004064abdaa56dbd0d231d281)]:
+  - @sanomalearning/slds-core@0.1.0
+
 ## 0.0.1
 
 ### Patch Changes